refactor: decouple onboarding repository
