+++ src/apps/server/onboardings/onboarding.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 
 import { OnboardingsController } from '@apps/server/onboardings/onboarding.controller';
 import { OnboardingsService } from '@apps/server/onboardings/onboarding.service';
-import { OnboardingRepository } from '@libs/modules/database/repositories/onboarding.repository';
+import { OnboardingRepository } from '@libs/modules/database/repositories/onboarding/onboarding.interface';
+import { OnboardingRepositoryImpl } from '@libs/modules/database/repositories/onboarding/onboarding.repository';
 
 @Module({
   controllers: [OnboardingsController],
-  providers: [OnboardingsService, OnboardingRepository],
+  providers: [OnboardingsService, { provide: OnboardingRepository, useClass: OnboardingRepositoryImpl }],
   exports: [OnboardingsService],
 })
 export class OnboardingsModule {}
