+++ src/libs/modules/database/repositories/onboarding/onboarding.interface.ts
@@ -1,0 +1,6 @@
+export interface OnboardingRepository {
+  findFirst;
+  update;
+}
+
+export const OnboardingRepository = Symbol('OnboardingRepository');
